pyproject.toml
[project]
name = "timm-cchess-reg"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.1.0",
    "coremltools>=8.0",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=10.0",
    "timm>=1.0.26",
    "torch>=2.5",
    "torchvision>=0.20",
    "transformers>=4.40",
]
